refactor(mesh): refactor multiscale mesh support
This changes the octree representation to be more explicit, and eliminates the reliance on being
able to split mesh fragments down to individual base chunks; instead, mesh fragments need only be
split to the next octree level.

This also adds support for more flexible vertex position representation, which will be needed for
the precomputed multiscale mesh format to be added in a later commit.

+++ src/neuroglancer/skeleton/backend.ts
@@ -148,6 +148,6 @@ export function decodeSkeletonVertexPositionsAndIndices(
   const meshData = decodeVertexPositionsAndIndices(
       /*verticesPerPrimitive=*/2, data, endianness, vertexByteOffset, numVertices,
     indexByteOffset, numEdges);
-  chunk.vertexPositions = meshData.vertexPositions;
+  chunk.vertexPositions = meshData.vertexPositions as Float32Array;
   chunk.indices = meshData.indices as Uint32Array;
 }
